moved sonarr to new base api, updated base apis
