+++ pyarr/__init__.py
@@ -1,6 +1,6 @@
 from .radarr import RadarrAPI
 from .request_api import RequestAPI
-from .sonarr_api import SonarrAPI
+from .sonarr import SonarrAPI
 from .readarr import ReadarrAPI
 
 __all__ = ["SonarrAPI", "RadarrAPI", "RequestAPI", "ReadarrAPI"]
